requirements.txt
absl-py==1.1.0
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==3.1.3
django-admin-interface==0.29.2
Django-Basic-CRUD-Generator==0.2
django-bootstrap==0.2.4
django-bootstrap-pagination==1.7.1
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-ckeditor-5==0.2.7
django-colorfield==0.11.0
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-filter==21.1
django-js-asset==2.0.0
django-libs==2.0.3
django-sslserver==0.22
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
pillow==10.4.0
PyJWT==1.7.1
python-slugify==8.0.4
pytz==2024.2
soupsieve==2.6
sqlparse==0.5.3
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==4.12.2
python-dateutil==2.8.2
python-docx==1.1.2
python-docx-replace==0.4.4